Organizing Modular Structure for Mutations
Still not quite following Vuex recommonded practices because I put my
actions file inside my service folder.  Vuex suggests that actions can
possibly call multiple mutations and should be global.
