+++ client/src/vuex/store.js
@@ -1,18 +1,11 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import messageMutations from './messages/mutations'
+import messages from './messages/mutations'
 
 Vue.use(Vuex)
 
-const state = {
-  messages: []
-}
-
-const mutations = {
-  ...messageMutations
-}
-
 export default new Vuex.Store({
-  state,
-  mutations
+  modules: {
+    messages
+  }
 })
