+++ server/src/hooks/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 // Add any common hooks you want to share across services in here.
-// 
+//
 // Below is an example of how a hook is written and exported. Please
 // see http://docs.feathersjs.com/hooks/readme.html for more details
 // on hooks.
@@ -9,5 +9,6 @@
 exports.myHook = function(options) {
   return function(hook) {
     console.log('My custom global hook ran. Feathers is awesome!');
+    return new Promise((resolve) => setTimeout(() => resolve() ,2000))
   };
 };
